Updates :- 1. Refactored individual task into separate files. 2. Updated Task Result structure via adding more detailed information.
