+++ internal/tasks/task_manager.go
@@ -42,6 +42,7 @@ func (tm *TaskManager) AddTask(task *Task) error {
 // StartTaskManager will initiate a scheduling job that will listen for taskQueue until
 // TaskManager is closed.
 func (tm *TaskManager) StartTaskManager() {
+	// TODO : Need to sleep incoming task from same host.bucket.scope.collection if such task from same host exists.
 	if tm.ctx.Err() != nil {
 		return
 	}
